Python-Password-Manager/requirements.txt
commonmark==0.9.1
mysql-connector-python==8.0.29
protobuf==3.20.1
pycryptodome==3.14.1
Pygments==2.12.0
pyperclip==1.8.2
rich==12.3.0
